fix(hono): zValidator wrapper doesn't typecheck with 0.7.4 (#2468)
Update the zValidator wrapper with the latest polymorphic/backwards compatible types from
`@hono/zod-validator` 0.7.4. Try hard to not rely on `any`. Type check the code with `tsc` during
linting to catch errors.

I'm not personally a fan of how `@hono/zod-validator` is typed. This may be a situation, however,
where TypeScript's verifier can't exhaustively check things because of the lack of higher-kinded
types. I think the solution is for them to provide overloads for v3 and v4 zod schemata and make the
 innards of the function very vague about types.

Signed-off-by: Alexandra Parker <alexandra.i.parker.civ@us.navy.mil>
Co-authored-by: Alexandra Parker <alexandra.i.parker.civ@us.navy.mil>